Python cleanup
- Define dependencies using poetry
- Update directory structure to follow standard Python practices
- add pylint and black
- add Makefile for building lambda package
- rework unit test to run all tests without needing to modify source
  code
