+++ aviatrix_ha/errors/exceptions.py
@@ -1,0 +1,5 @@
+""" Module to hold all errors"""
+
+
+class AvxError(Exception):
+    """Error class for Aviatrix exceptions"""
