+++ aviatrix_ha/version.py
@@ -1,0 +1,3 @@
+""" Version information"""
+
+VERSION = "2.02"
