+++ pyproject.toml
@@ -1,0 +1,22 @@
+[tool.poetry]
+name = "aviatrix_ha"
+version = "2.0.2"
+description = "Aviatrix Controller HA Lambda functions"
+authors = ["Aviatrix <support@aviatrix.com>"]
+license = "Apache 2.0"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+requests = "^2.31.0"
+urllib3 = "^2.2.1"
+
+[tool.poetry.group.dev.dependencies]
+responses = "^0.25.0"
+pytest = "^8.1.1"
+boto3 = "^1.34.64"
+pylint = "^3.1.0"
+black = "^24.3.0"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
