--- src/errors/exceptions.py
@@ -1,5 +1,0 @@
-""" Module to hold all errors"""
-
-
-class AvxError(Exception):
-    """ Error class for Aviatrix exceptions"""
