--- src/version.py
@@ -1,2 +1,0 @@
-""" Version information"""
-VERSION = "2.02"
